+++ scripts/build.js
@@ -168,6 +168,7 @@ await (async () => {
         `
           :host {
             display: block;
+            flex-shrink: 0;
             ${
               path.startsWith("24")
                 ? "width: 1.50rem; height: 1.50rem;"