+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.4.0] - 2024-01-05
+
 ### Changed
 
 - Components css now includes width, height & `flex-shrink: 0`.
@@ -40,7 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/malobre/heroicons-wc/compare/v0.3.1...HEAD
+[unreleased]: https://github.com/malobre/heroicons-wc/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/malobre/heroicons-wc/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/malobre/heroicons-wc/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/malobre/heroicons-wc/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/malobre/heroicons-wc/compare/v0.1.0...v0.2.0